feature: Reworked the Person TopBilled card component
