+++ src/components/Cards/Entertainment/Recommendation/Recommendation.jsx
@@ -7,9 +7,24 @@ import Image from '../../../Core/Image';
 
 import replacesSpacesWith from '../../../../utils/formatters/replaceSpacesWith';
 
-const Recommendation = ({ title, releaseDate, image, rating }) => (
-	<div className='cursor-pointer max-w-[300px] min-w-[300px] p-4'>
-		<div className='relative group'>
+const Recommendation = ({ title, releaseDate, image, rating, onClick, onKeyDown }) => (
+	<div className='cursor-pointer max-w-[300px] min-w-[300px]'>
+		<div
+			className='relative group'
+			onClick={(event) => {
+				if (onClick) {
+					onClick(event);
+				}
+			}}
+			type='button'
+			tabIndex={0}
+			onKeyDown={(event) => {
+				if (onKeyDown) {
+					onKeyDown(event);
+				}
+			}}
+			role='button'
+		>
 			<Image
 				width='100%'
 				height='100%'
@@ -35,14 +50,16 @@ Recommendation.propTypes = {
 	title: PropTypes.string,
 	releaseDate: PropTypes.string,
 	image: PropTypes.string,
-	rating: PropTypes.number
+	rating: PropTypes.number,
+	onClick: PropTypes.func
 };
 
 Recommendation.defaultProps = {
 	title: 'Ant Man and The Wasp',
 	releaseDate: '07/04/2020',
 	image: 'https://image.tmdb.org/t/p/original/6P3c80EOm7BodndGBUAJHHsHKrp.jpg',
-	rating: 70
+	rating: 70,
+	onClick: null
 };
 
 export default Recommendation;
